third_party/WebKit/LayoutTests/fast/forms/file/input-file-value-expected.txt

This tests the value attribute in file input forms

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS file.value is "C:\\fakepath\\foo.txt"
PASS file.files.length is 1
PASS file.value = 'foo' threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
PASS file.value is "C:\\fakepath\\foo.txt"
PASS file.files.length is 1
PASS file.value is ""
PASS file.files.length is 0
PASS successfullyParsed is true

TEST COMPLETE
